requirements.txt
nose~=1.3.7
sphinx~=7.1.2
coloredlogs~=15.0.1
openpyxl~=3.0.0
setuptools~=68.2.0
pandas~=2.0.3
tabulate~=0.9.0
pip~=23.2.1
wheel~=0.41.2
docutils~=0.20.1
Jinja2~=3.1.2
pytz~=2024.2
zipp~=3.8.0
babel~=2.16.0
MarkupSafe~=2.1.1
pyreadline3~=3.5.4
pyreadline3~=3.5.4
packaging~=21.3
imagesize~=1.4.1
colorama~=0.4.6
requests~=2.32.3
urllib3~=2.2.3
snowballstemmer~=2.2.0
importlib-metadata~=4.11.3
Pygments~=2.12.0
certifi~=2024.8.30
et-xmlfile~=1.0.1
numpy~=1.24.4
idna~=3.10
charset-normalizer~=3.4.0
alabaster~=0.7.13
humanfriendly~=10.0
platformdirs~=4.3.6
mdutils~=1.6.0
SQLAlchemy~=2.0.36
pillow~=10.4.0
lxml~=5.3.0
python-docx~=1.1.2